修改了大量代码，先提交下

+++ lib/src/main/java/simple/media/player/presenter/ControllerDurationPresenter.java
@@ -118,8 +118,8 @@ public class ControllerDurationPresenter {
             return;
         }
 
-        int duration = model.getSimpleMediaPlayer().getDurationInMs();
-        int current = model.getSimpleMediaPlayer().getCurrentPositionInMs();
+        int duration = (int) model.getSimpleMediaPlayer().getDurationInMs();
+        int current = (int) model.getSimpleMediaPlayer().getCurrentPositionInMs();
 
         int pc = duration == 0 ? 0 : (int) (current * 1.0f / duration * 100);
         //Log.i(SimpleMediaPlayer.TAG, "cur:" + current + ",dur:" + duration + "," + pc + "%");
